+++ unit1bGame/Item.h
@@ -5,13 +5,14 @@
 class Item {
 	constants::ITEM_TYPE type;
 	string displayName;
-	vector<int> instanceData;
+	//throw item type specific data in here
+	vector<double> instanceData;
 public:
 	Item(constants::ITEM_TYPE newType = constants::nullItemType, string newDisplayName = "UNINITIALIZED ITEM") : type(newType), displayName(newDisplayName) {};
 	//downward casting depends on this
 	virtual ~Item() {};
 
-	vector<int>& getInstanceData();
+	vector<double>& getInstanceData();
 
 	constants::ITEM_TYPE getType();
 	void setType(constants::ITEM_TYPE newType);
